+++ packages/profiler/src/interfaces/powermetrics/PowerMetricsSensorInterface.ts
@@ -106,7 +106,10 @@ export class PowerMetricsSensorInterface extends BaseSensorInterface {
 		}
 	}
 
-	async readSensorValues(pid: number): Promise<MetricsDataCollection> {
+	async readSensorValues(pid: number): Promise<MetricsDataCollection | undefined> {
+		if (!await this.couldBeExecuted()) {
+			return undefined
+		}
 		let tries = 0
 		while (this.isRunning() && tries < 10) {
 			console.error(`Cannot read sensor values, wait for process to exit: ${tries + 1}, try again after 1 second`)
@@ -157,6 +160,9 @@ export class PowerMetricsSensorInterface extends BaseSensorInterface {
 	}
 
 	async startProfiling() {
+		if (!await this.couldBeExecuted()) {
+			return
+		}
 		if (fs.existsSync(this._options.outputFilePath)) {
 			fs.unlinkSync(this._options.outputFilePath) // remove output file to ensure clean measurements
 		}
@@ -183,6 +189,9 @@ export class PowerMetricsSensorInterface extends BaseSensorInterface {
 	}
 
 	async stopProfiling() {
+		if (!await this.couldBeExecuted()) {
+			return
+		}
 		if (this._childProcess === undefined) {
 			return
 		}
